+++ ot_simple_rest/handlers/jobs/db_connector.py
@@ -15,11 +15,11 @@ class PostgresConnector(PGConnector):
     def __init(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
-    def check_cache(self, *, original_spl, tws, twf, field_extraction, preview):
+    def check_cache(self, *, original_otl, tws, twf, field_extraction, preview):
         cache_id = creating_date = None
         query_str = "SELECT id, extract(epoch from creating_date) FROM cachesdl WHERE expiring_date >= " \
-                    "CURRENT_TIMESTAMP AND original_spl=%s AND tws=%s AND twf=%s AND field_extraction=%s AND preview=%s;"
-        stm_tuple = (original_spl, tws, twf, field_extraction, preview)
+                    "CURRENT_TIMESTAMP AND original_otl=%s AND tws=%s AND twf=%s AND field_extraction=%s AND preview=%s;"
+        stm_tuple = (original_otl, tws, twf, field_extraction, preview)
         self.logger.info(query_str % stm_tuple)
 
         cache_data = self.execute_query(query_str, params=stm_tuple, fetchall=True)
@@ -27,11 +27,11 @@ class PostgresConnector(PGConnector):
             cache_id, creating_date = cache_data
         return cache_id, creating_date
 
-    def check_running(self, *, original_spl, tws, twf, field_extraction, preview):
+    def check_running(self, *, original_otl, tws, twf, field_extraction, preview):
         job_id = creating_date = None
-        query_str = "SELECT id, extract(epoch from creating_date) FROM splqueries WHERE status = 'running' " \
-                    "AND original_spl=%s AND tws=%s AND twf=%s AND field_extraction=%s AND preview=%s;"
-        stm_tuple = (original_spl, tws, twf, field_extraction, preview)
+        query_str = "SELECT id, extract(epoch from creating_date) FROM OTLQueries WHERE status = 'running' " \
+                    "AND original_otl=%s AND tws=%s AND twf=%s AND field_extraction=%s AND preview=%s;"
+        stm_tuple = (original_otl, tws, twf, field_extraction, preview)
         self.logger.info(query_str % stm_tuple)
 
         job_data = self.execute_query(query_str, params=stm_tuple, fetchall=True)
@@ -49,7 +49,7 @@ class PostgresConnector(PGConnector):
 
     def add_job(self, *, search, subsearches, tws, twf, cache_ttl, username, field_extraction, preview):
         job_id = creating_date = None
-        query_str = "INSERT INTO splqueries (original_spl, service_spl, subsearches, tws, twf, cache_ttl, username, " \
+        query_str = "INSERT INTO OTLQueries (original_otl, service_otl, subsearches, tws, twf, cache_ttl, username, " \
                     "field_extraction, preview) VALUES (%s,%s,%s,%s,%s,%s,%s,%s,%s) RETURNING id, " \
                     "extract(epoch from creating_date);"
         stm_tuple = (search[0], search[1], subsearches, tws, twf, cache_ttl, username, field_extraction, preview)
@@ -60,17 +60,17 @@ class PostgresConnector(PGConnector):
             job_id, creating_date = job_data
         return job_id, creating_date
 
-    def add_sid(self, *, sid, remote_ip, original_spl):
-        query_str = "INSERT INTO SplunkSIDs (sid, src_ip, spl) VALUES (%s, %s, %s);"
-        stm_tuple = (sid, remote_ip, original_spl)
+    def add_sid(self, *, sid, remote_ip, original_otl):
+        query_str = "INSERT INTO GUISIDs (sid, src_ip, spl) VALUES (%s, %s, %s);"
+        stm_tuple = (sid, remote_ip, original_otl)
         self.logger.info(query_str % stm_tuple)
         self.execute_query(query_str, params=stm_tuple, with_commit=True, with_fetch=False)
 
-    def add_external_job(self, *, original_spl, service_spl, tws, twf, cache_ttl, username, status):
+    def add_external_job(self, *, original_otl, service_otl, tws, twf, cache_ttl, username, status):
         cache_id = creating_date = None
-        query_str = "INSERT INTO splqueries (original_spl, service_spl, tws, twf, cache_ttl, username, status) " \
+        query_str = "INSERT INTO OTLQueries (original_otl, service_otl, tws, twf, cache_ttl, username, status) " \
                     "VALUES (%s,%s,%s,%s,%s,%s,%s) RETURNING id, extract(epoch from creating_date);"
-        stm_tuple = (original_spl, service_spl, tws, twf, cache_ttl, username, status)
+        stm_tuple = (original_otl, service_otl, tws, twf, cache_ttl, username, status)
         self.logger.debug(query_str % stm_tuple)
 
         cache_data = self.execute_query(query_str, params=stm_tuple, with_commit=True)
@@ -78,10 +78,10 @@ class PostgresConnector(PGConnector):
             cache_id, creating_date = cache_data
         return cache_id, creating_date
 
-    def add_to_cache(self, *, original_spl, tws, twf, cache_id, expiring_date):
-        query_str = "INSERT INTO CachesDL (original_spl, tws, twf, id, expiring_date) " \
+    def add_to_cache(self, *, original_otl, tws, twf, cache_id, expiring_date):
+        query_str = "INSERT INTO CachesDL (original_otl, tws, twf, id, expiring_date) " \
                     "VALUES(%s, %s, %s, %s, to_timestamp(extract(epoch from now()) + %s));"
-        stm_tuple = (original_spl, tws, twf, cache_id, expiring_date)
+        stm_tuple = (original_otl, tws, twf, cache_id, expiring_date)
         self.logger.debug(query_str % stm_tuple)
         self.execute_query(query_str, params=stm_tuple, with_commit=True, with_fetch=False)
 
@@ -90,7 +90,7 @@ class PostgresConnector(PGConnector):
         return self.execute_query(query_str, params=(datamodel_name,), fetchall=True)
 
     def get_spl(self, sid, src_ip):
-        query_str = "SELECT spl FROM SplunkSIDs WHERE sid=%s AND src_ip=%s;"
+        query_str = "SELECT spl FROM GUISIDs WHERE sid=%s AND src_ip=%s;"
         return self.execute_query(query_str, params=(sid, src_ip,), fetchall=True)
 
     def check_dispatcher_status(self):
@@ -101,12 +101,12 @@ class PostgresConnector(PGConnector):
         if time_delta:
             return time_delta[0]
 
-    def check_job_status(self, *, original_spl, tws, twf, field_extraction, preview):
-        query_str = "SELECT splqueries.id, splqueries.status, cachesdl.expiring_date, splqueries.msg " \
-                    "FROM splqueries LEFT JOIN cachesdl ON splqueries.id = cachesdl.id " \
-                    "WHERE splqueries.original_spl=%s AND splqueries.tws=%s AND splqueries.twf=%s " \
-                    "AND splqueries.field_extraction=%s AND splqueries.preview=%s ORDER BY splqueries.id DESC LIMIT 1;"
-        stm_tuple = (original_spl, tws, twf, field_extraction, preview)
+    def check_job_status(self, *, original_otl, tws, twf, field_extraction, preview):
+        query_str = "SELECT OTLQueries.id, OTLQueries.status, cachesdl.expiring_date, OTLQueries.msg " \
+                    "FROM OTLQueries LEFT JOIN cachesdl ON OTLQueries.id = cachesdl.id " \
+                    "WHERE OTLQueries.original_otl=%s AND OTLQueries.tws=%s AND OTLQueries.twf=%s " \
+                    "AND OTLQueries.field_extraction=%s AND OTLQueries.preview=%s ORDER BY OTLQueries.id DESC LIMIT 1;"
+        stm_tuple = (original_otl, tws, twf, field_extraction, preview)
         self.logger.info(query_str % stm_tuple)
 
         job_data = self.execute_query(query_str, params=stm_tuple)