+++ ot_simple_rest/handlers/jobs/loadjob.py
@@ -18,7 +18,7 @@ class LoadJob(tornado.web.RequestHandler):
     This handler helps OT.Simple Splunk app JobLoader to check Job's status and then to download results from ramcache.
 
     1. Remove OT.Simple Splunk app service data from SPL query.
-    2. Get Job's status based on (original_spl, tws, twf) parameters.
+    2. Get Job's status based on (original_otl, tws, twf) parameters.
     3. Check Job's status and return it to OT.Simple Splunk app if it is not still ready.
     4. Load results of Job from cache for transcending.
     5. Return Job's status or results.