+++ sql/eva.sql
@@ -75,4 +75,5 @@ INSERT INTO "user" (name, password) VALUES ('admin', '$2b$12$ODxOO2wd6vRy2wT4euC
 INSERT INTO permission (name) VALUES ('admin_all');
 INSERT INTO role (name) VALUES ('admin');
 INSERT INTO user_role (user_id, role_id) VALUES (1, 1);
-INSERT INTO role_permission (role_id, permission_id) VALUES (1, 1);
+INSERT INTO index (name) VALUES ('*');
+INSERT INTO index_group VALUES (1, 1);